requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
arrow==1.2.2
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.4
config==0.5.1
cryptography==36.0.1
ddt==1.5.0
et-xmlfile==1.1.0
Faker==8.13.2
Flask==2.0.3
h11==0.13.0
idna==3.2
iniconfig==1.1.1
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
openpyxl==3.0.7
outcome==1.1.0
packaging==21.0
paramiko==2.9.2
pluggy==1.0.0
py==1.10.0
PyAutoIt==0.6.5
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pypiwin32==223
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sshtunnel==0.4.0
text-unidecode==1.3
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.6
Werkzeug==2.0.3
wsproto==1.1.0
xlrd==1.2.0
XMind==1.2.0
xmind2testcase==1.5.0
